refactor: Car의 이름을 Name으로 래핑한다
