+++ src/main/java/study/racinggame/domain/car/Car.java
@@ -1,17 +1,11 @@
 package study.racinggame.domain.car;
 
 public class Car {
-    public static final int LENGTH_LIMIT = 5;
-
-    private final String name;
+    private final Name name;
     private int position;
 
     public Car(String name, int position) {
-        if (name.length() < 1 || name.length() > LENGTH_LIMIT) {
-            throw new IllegalArgumentException("자동차 이름은 5자를 초과할 수 없습니다.");
-        }
-
-        this.name = name;
+        this.name = new Name(name);
         this.position = position;
     }
 
@@ -25,15 +19,15 @@ public class Car {
         }
     }
 
+    public String name() {
+        return name.getValue();
+    }
+
     public boolean isWinner(int max) {
         return this.position == max;
     }
 
     public int getPosition() {
         return position;
-    }
-
-    public String getName() {
-        return name;
     }
 }
