+++ src/main/java/study/racinggame/view/OutputView.java
@@ -12,16 +12,18 @@ public class OutputView {
 
     public static void printPosition(Cars cars) {
         cars.getCars().forEach(car -> {
-            System.out.print(car.getName() + ": ");
+            System.out.print(car.name() + ": ");
+
             for (int i = 0; i < car.getPosition(); i++) {
                 System.out.print("-");
             }
+
             System.out.println();
         });
     }
 
     public static void printWinners(List<Car> cars) {
         System.out.print("\n우승자: ");
-        cars.forEach(car -> System.out.print(car.getName() + " "));
+        cars.forEach(car -> System.out.print(car.name() + " "));
     }
 }
